:truck: Move config-files
