+++ aksel.nav.no/website/app/(routes)/routing-config.ts
