+++ aksel.nav.no/website/tsconfig.json
@@ -19,7 +19,7 @@
       "node",
       "umami",
       "@navikt/ds-react/types/theme",
-      "./app/aksel-theme.d.ts"
+      "./app/theme.d.ts"
     ],
     "paths": {
       "@/sb-util": ["./.storybook/utils.tsx"],
